Cleaned up code, restructured
