+++ src/styles.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Open+Sans:wght@300;400&display=swap');
+
 html{line-height:1.15;-ms-text-size-adjust:100%;-webkit-text-size-adjust:100%}body{margin:0}article,aside,footer,header,nav,section{display:block}h1{font-size:2em;margin:.67em 0}figcaption,figure,main{display:block}figure{margin:1em 40px}hr{box-sizing:content-box;height:0;overflow:visible}pre{font-family:monospace,monospace;font-size:1em}a{background-color:transparent;-webkit-text-decoration-skip:objects}abbr[title]{border-bottom:none;text-decoration:underline;text-decoration:underline dotted}b,strong{font-weight:inherit}b,strong{font-weight:bolder}code,kbd,samp{font-family:monospace,monospace;font-size:1em}dfn{font-style:italic}mark{background-color:#ff0;color:#000}small{font-size:80%}sub,sup{font-size:75%;line-height:0;position:relative;vertical-align:baseline}sub{bottom:-.25em}sup{top:-.5em}audio,video{display:inline-block}audio:not([controls]){display:none;height:0}img{border-style:none}svg:not(:root){overflow:hidden}button,input,optgroup,select,textarea{font-family:sans-serif;font-size:100%;line-height:1.15;margin:0}button,input{overflow:visible}button,select{text-transform:none}button,html [type=button],[type=reset],[type=submit]{-webkit-appearance:button}button::-moz-focus-inner,[type=button]::-moz-focus-inner,[type=reset]::-moz-focus-inner,[type=submit]::-moz-focus-inner{border-style:none;padding:0}button:-moz-focusring,[type=button]:-moz-focusring,[type=reset]:-moz-focusring,[type=submit]:-moz-focusring{outline:1px dotted ButtonText}fieldset{padding:.35em .75em .625em}legend{box-sizing:border-box;color:inherit;display:table;max-width:100%;padding:0;white-space:normal}progress{display:inline-block;vertical-align:baseline}textarea{overflow:auto}[type=checkbox],[type=radio]{box-sizing:border-box;padding:0}[type=number]::-webkit-inner-spin-button,[type=number]::-webkit-outer-spin-button{height:auto}[type=search]{-webkit-appearance:textfield;outline-offset:-2px}[type=search]::-webkit-search-cancel-button,[type=search]::-webkit-search-decoration{-webkit-appearance:none}::-webkit-file-upload-button{-webkit-appearance:button;font:inherit}details,menu{display:block}summary{display:list-item}canvas{display:inline-block}template{display:none}[hidden]{display:none}
 
 .clearfix:after {visibility: hidden; display: block; font-size: 0; content: ''; clear: both; height: 0; }
@@ -12,6 +14,7 @@ html {
     background-color: #000;
     color: #FFF;
     font-size: 62.5%;
+    font-family: 'Open Sans', sans-serif;
 }
 
 body {
@@ -24,17 +27,20 @@ body {
 
 header {
     text-align: center;
-    font-size: 1.2rem;
+    font-size: 1rem;
     text-transform: uppercase;
     margin-bottom: 10px;
     height: 15px;
+    font-weight: 700;
 }
 
 .stopwatchUpper {
-    height: calc((100vh - 45px) / 2) ;
+    height: calc((100vh - 45px) / 2);
     display: flex;
     flex-direction: column;
     justify-content: space-between;
+    border-bottom: 1px solid #212121;
+    padding-bottom: 20px;
 }
 
 .stopwatchLower {
@@ -54,6 +60,7 @@ header {
     display: flex;
     justify-content: center;
     align-items: center;
+    font-weight: 300;
 }
 
 .stopwatchUpper .stopwatchButtonContainer {
@@ -69,6 +76,7 @@ header {
     border: none;
     background-color: #333333;
     color: #FFF;
+    font-family: 'Open Sans', sans-serif;
 }
 
 .stopwatchUpper .stopwatchButtonContainer button.start {
@@ -82,7 +90,8 @@ header {
 }
 
 .stopwatchLower ul {
-    padding-left: 0;
+    margin: 0;
+    padding: 0;
 }
 
 .stopwatchLower li {
